Add Simple JWT Auth. Improve Reservation & Ticket functionality. Remake user as a separate application
Added JWT based authentication. Now access to the endpoints is only available for authenticated users.
Remade Reservation functionality so that user can choose from tickets that are not reserved.
Added validation so that, user can not reserve the same ticket twice, and ticket's row and seats must be in certain range to be valid.
Made user model as a separate application as it is more convinient and efficient to work with.